Splitting controllers
- Splitting controllers, view and manage
